requirements.txt
django==3.2.12
psycopg2-binary==2.9.3
